refactor(media): name the client quality control explicitly
